requirements.txt
peewee
pymysql
passlib
grpcio
grpcio-tools
protobuf
loguru
python-consul2
redis
jaeger_client
pybreaker
ratelimit